tracking with own map seems to start working with simulated data

+++ source/tracking/predict.m
@@ -19,13 +19,14 @@ function particles = predict(particles, deltaT_global)
 % the longer we wait between events, the longer the camera probably moved
 % also allow some error even when the camera didn't move at all
 sigma = 0.00001 + 0.00002 * deltaT_global;
+% sigma = 0.0000001 + 0.0000002 * deltaT_global;
 
 % TODO: implement perturbation in tangent space
 %n = randn(1,3) * sigma; % zero mean
 
 % TODO: this would probably work as well, right? (I don't know enough about
 % Lie algebra and exponential maps)
-particles(:, 2:end) = particles(:, 2:end) + sigma * randn(size(particles)-[0,1]);
+particles(:, 2:end) = particles(:, 2:end) + double(sigma) * randn(size(particles)-[0,1]);
 
 disp(['predict: deltaT = ' num2str(deltaT_global)]);
